Add files via upload
This project demonstrates the application of data analysis and machine learning techniques using a dataset (data.csv). The goal of the project is to explore and clean the data, build a machine learning model (Random Forest Regressor), and fine-tune the model using Grid Search Cross Validation. We will also explore the data through visualizations such as pair plots, bar plots, and histograms to understand its distribution and relationships.